Ordering is now sort of tested, broke out length() to prove it.
However it might be revealing a problem in PPI. At least I don't see any
reason why the content is being found correctly in order to alter it
once, but not the second time.

Worst case I pull it out of PPI and figure out the unswizzling logic.

Here-docs and regexes could be problematic though.