Reordering works better when calling them dependencies...
But there is an 'XXX' here. Even though the order appears to be right, I
can't stop '.' 'elems' from being parsed incorrectly. I fear that
something might be going on with PPI because I've seen similar issues
while working with this code.

Going to try an experiment here for a while...